Move TradePairSorter to a new file
